core: RPMB FS: use ADD_OVERFLOW() in write function
rpmb_fs_write_primitive() could potentially allow writes outside of
legitimate memory, due to a lack of overflow check when computing
addresses.

'pos' is under control of the TA through the seek function and could
potentially be made such that newsize is too small to receive the data
copied in the allocate/read/copy/update branch of the code.

Use the ADD_OVERFLOW() macro to detect this situation and return an
error.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reported by: Bastien Simondi <bsimondi@netflix.com> [1.9]
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>