Replace ta_head.entry with elf entry
Prior to this patch the entry function of the TA was stored in ta_head
which is located in a read-only section of the TA. This results in the
linker emitting a relocation modifying a read-only section. This is a
problem if the read-only section is mapped read-only while relocations
are performed. To avoid this problematic relocation the ta_head.entry
is removed and the ELF entry point is used instead.

Note that this change is incompatible with previously compiled TAs.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>