core: change tee_pager_add_uta_area() arguments
Simplifies tee_pager_add_uta_area() by taking a pointer to a struct fobj
instead instead of a size. The return value is changed to a TEE_Result
to harmonize better with other functions.

vm_map() is changed to expect a mobj with an assigned fobj when paging
is enabled. This requires that the pager allocations done with
mobj_paged_alloc() are replaced with fobj_rw_paged_alloc().

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>