Add TA entry point function: __ta_entry()
The ELF entry point of a TA is currently not used by OP-TEE. However,
there is some work ongoing in the ELF loader which will require the
entry point to execute __utee_entry [1].

On the other hand, the Clang linker will warn if it links an executable
that has no _start symbol and no entry point is specified:

 ld.lld: warning: cannot find entry symbol _start; defaulting to 0x20

This commits adds the entry point already. Note that the __utee_entry
symbol may be undefined in a TA in case libutee is built as a shared
library (CFG_ULIBS_SHARED=y). Therefore, add a wrapper function to
avoid this issue.

Link: [1] https://github.com/OP-TEE/optee_os/pull/2801
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>