libutils: strip __builtin_ prefix from implementations
Our implementations of the <ctypes.h> functions (isalnum(), isalpha(),
etc.) are defined with the __builtin_ prefix. This is not needed and
is arguably wrong. Indeed, we want the functions in libutils to be
present without the prefix. The purpose of the #defines in
lib/libutils/isoc/include/ctype.h is to tell the compiler that it is
allowed to recognize the functions and optimize them; the
-ffreestanding flag would prevent this otherwise.

This __builtin_ prefix in the .c files makes things more complicated to
support Clang. Therefore, remove it.

The libutils binary is identical with and without this patch.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>